code review after article part 2

+++ buzzy.sh
@@ -15,8 +15,12 @@ case "$1" in
     ;;
   --clean-deep)
     ./gradlew clean &&
-    rm -rf buzzy-webapp/dist &&
-    rm -rf buzzy-webapp/node_modules
+    read -p "Are you sure you want to remove dist and node_modules? (Y/n): " -r confirm
+    confirm=${confirm:-Y}
+    if [[ $confirm =~ ^[Yy]$ ]]; then
+      rm -rf buzzy-webapp/dist &&
+      rm -rf buzzy-webapp/node_modules
+    fi
     ;;
   --build)
     ./gradlew clean build
@@ -29,23 +33,28 @@ case "$1" in
     docker compose up -d --build buzzy_webapp
     ;;
   --start-dev)
-    docker compose up -d --build buzzy_sso_db buzzy_api_db buzzy_redis buzzy_kafka buzzy_kafka_zookeeper
+    docker compose up -d --build buzzy_sso_db buzzy_api_db buzzy_redis buzzy_kafka_ui buzzy_kafka
     ;;
   --down)
     docker compose down
     ;;
   --down-clean)
-    read -p "Are you sure you want to remove the containers and its data? (y/n): " confirm
-    if [ "$confirm" == "y" ]; then
+    read -p "Are you sure you want to remove the containers and its data? (Y/n): " -r confirm
+    confirm=${confirm:-Y}
+    if [[ $confirm =~ ^[Yy]$ ]]; then
       docker compose down &&
-      docker volume rm buzzy_buzzy_sso_db_data buzzy_buzzy_redis_data buzzy_buzzy_kafka_zookeeper_data buzzy_buzzy_api_db_data
+      docker volume rm buzzy_buzzy_sso_db_data buzzy_buzzy_redis_data buzzy_buzzy_kafka_data buzzy_buzzy_api_db_data
     fi
     ;;
   --down-clean-deep)
-    read -p "Are you sure you want to remove the containers and its data, build, dist and node_modules? (y/n): " confirm
-    if [ "$confirm" == "y" ]; then
+    read -p "Are you sure you want to remove the containers and its data, build, dist and node_modules? (Y/n): " -r confirm
+    confirm=${confirm:-Y}
+    if [[ $confirm =~ ^[Yy]$ ]]; then
       docker compose down &&
-      docker volume rm buzzy_buzzy_sso_db_data buzzy_buzzy_redis_data buzzy_buzzy_kafka_zookeeper_data buzzy_buzzy_api_db_data
+    ./gradlew clean &&
+    rm -rf buzzy-webapp/dist &&
+    rm -rf buzzy-webapp/node_modules &&
+    docker volume rm buzzy_buzzy_sso_db_data buzzy_buzzy_redis_data buzzy_buzzy_kafka_data buzzy_buzzy_api_db_data
     fi
     ;;
   *)
